requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
supabase==1.0.3
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-core==2.10.1
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3